+++ Source/ImoutoRebirth.Navigator/ImoutoRebirth.Navigator/Services/Collections/SourceFolderService.cs
@@ -21,7 +21,9 @@ internal class SourceFolderService : ISourceFolderService
                 ShouldCheckHashFromName: x.ShouldCheckHashFromName,
                 ShouldCreateTagsFromSubfolders: x.ShouldCreateTagsFromSubfolders,
                 ShouldAddTagFromFilename: x.ShouldAddTagFromFilename,
-                SupportedExtensions: x.SupportedExtensions ?? Array.Empty<string>()
+                SupportedExtensions: x.SupportedExtensions ?? Array.Empty<string>(),
+                IsWebhookUploadEnabled: x.IsWebhookUploadEnabled,
+                WebhookUploadUrl: x.WebhookUploadUrl
             ))
             .ToArray();
     }
@@ -30,12 +32,14 @@ internal class SourceFolderService : ISourceFolderService
     {
         var id = await _collectionsClient.AddSourceFolderAsync(new(
             sourceFolder.CollectionId,
+            sourceFolder.IsWebhookUploadEnabled,
             sourceFolder.Path,
             sourceFolder.ShouldAddTagFromFilename,
             sourceFolder.ShouldCheckFormat,
             sourceFolder.ShouldCheckHashFromName,
             sourceFolder.ShouldCreateTagsFromSubfolders,
-            sourceFolder.SupportedExtensions));
+            sourceFolder.SupportedExtensions,
+            sourceFolder.WebhookUploadUrl));
 
         return sourceFolder with { Id = id };
     }
@@ -43,17 +47,19 @@ internal class SourceFolderService : ISourceFolderService
     public async Task UpdateSourceFolderAsync(SourceFolder sourceFolder)
     {
         if (!sourceFolder.Id.HasValue)
-            throw new ArgumentException("Can't update new collection", nameof(sourceFolder));
+            throw new ArgumentException(nameof(sourceFolder));
 
         await _collectionsClient.UpdateSourceFolderAsync(new(
             sourceFolder.CollectionId,
+            sourceFolder.IsWebhookUploadEnabled,
             sourceFolder.Path,
             sourceFolder.ShouldAddTagFromFilename,
             sourceFolder.ShouldCheckFormat,
             sourceFolder.ShouldCheckHashFromName,
             sourceFolder.ShouldCreateTagsFromSubfolders,
             sourceFolder.Id.Value,
-            sourceFolder.SupportedExtensions));
+            sourceFolder.SupportedExtensions,
+            sourceFolder.WebhookUploadUrl));
     }
 
     public async Task DeleteSourceFolderAsync(Guid collectionId, Guid sourceFolderId)