+++ Source/ImoutoRebirth.Navigator/ImoutoRebirth.Navigator/ViewModel/SettingsSlice/SourceFolderVM.cs
@@ -1,5 +1,7 @@
 ﻿using System.Collections.ObjectModel;
+using System.ComponentModel.DataAnnotations;
 using CommunityToolkit.Mvvm.ComponentModel;
+using ImoutoRebirth.Common.WPF.ValidationAttributes;
 
 namespace ImoutoRebirth.Navigator.ViewModel.SettingsSlice;
 
@@ -17,6 +19,15 @@ internal partial class SourceFolderVM : FolderVM
     [ObservableProperty]
     private bool _addTagFromFileName;
 
+    [ObservableProperty]
+    private bool _isWebhookUploadEnabled;
+
+    [ObservableProperty]
+    [NotifyCanExecuteChangedFor(nameof(SaveCommand))]
+    [NotifyDataErrorInfo]
+    [CustomValidation(typeof(SourceFolderVM), nameof(ValidateWebhookUploadUrl))]
+    private string? _webhookUploadUrl;
+
     public ObservableCollection<string> SupportedExtensionsRaw { get; }
 
     public SourceFolderVM(
@@ -26,13 +37,17 @@ internal partial class SourceFolderVM : FolderVM
         bool checkNameHash, 
         IEnumerable<string>? extensions, 
         bool tagsFromSubfolder, 
-        bool addTagFromFilename) 
+        bool addTagFromFilename,
+        bool isWebhookUploadEnabled,
+        string? webhookUploadUrl)
         : base(id, path)
     {
         CheckFormat = checkFormat;
         CheckNameHash = checkNameHash;
         TagsFromSubfolder = tagsFromSubfolder;
         AddTagFromFileName = addTagFromFilename;
+        IsWebhookUploadEnabled = isWebhookUploadEnabled;
+        WebhookUploadUrl = webhookUploadUrl;
 
         SupportedExtensionsRaw = extensions != null
             ? new ObservableCollection<string>(extensions) 
@@ -58,5 +73,18 @@ internal partial class SourceFolderVM : FolderVM
                 // ignore
             }
         }
+    }
+
+    public static ValidationResult ValidateWebhookUploadUrl(string? webhookUploadUrl, ValidationContext context)
+    {
+        var instance = (SourceFolderVM)context.ObjectInstance;
+
+        if (!instance.IsWebhookUploadEnabled)
+            return ValidationResult.Success!;
+
+        if (string.IsNullOrWhiteSpace(webhookUploadUrl))
+            return new("The WebhookUploadUrl field is required when WebhookUpload is enabled.");
+
+        return ValidationResult.Success!;
     }
 }