+++ Source/ImoutoRebirth.Room/ImoutoRebirth.Room.Application/ServiceCollectionExtensions.cs
@@ -19,8 +19,6 @@ public static class ServiceCollectionExtensions
         });
         services.AddLoggingBehavior();
         services.AddTransactionBehavior();
-        
-        services.AddSingleton<IImoutoPicsUploaderRepository, ImoutoPicsUploaderRepository>();
 
         return services;
     }