+++ Source/ImoutoRebirth.Room/ImoutoRebirth.Room.Application/Services/IImoutoPicsUploader.cs
@@ -1,6 +1,0 @@
-namespace ImoutoRebirth.Room.Application.Services;
-
-public interface IImoutoPicsUploader
-{
-    Task UploadFile(string path);
-}