+++ Source/ImoutoRebirth.Room/ImoutoRebirth.Room.Application/Services/IWebhookUploader.cs
@@ -1,0 +1,6 @@
+namespace ImoutoRebirth.Room.Application.Services;
+
+public interface IWebhookUploader
+{
+    Task UploadFile(string path, string uploadUrl);
+}