+++ Source/ImoutoRebirth.Room/ImoutoRebirth.Room.IntegrationTests/SourceFoldersTests.cs
@@ -34,7 +34,9 @@ public class SourceFoldersTests
             true,
             true,
             true,
-            new[] {".jpg", ".png", ".gif", ".webp", ".bmp", ".jpeg", ".tiff", ".tif", ".jfif"});
+            [".jpg", ".png", ".gif", ".webp", ".bmp", ".jpeg", ".tiff", ".tif", ".jfif"],
+            false,
+            null);
         
         var result =await httpClient.PostAsJsonAsync("/collections/source-folders", command);
         var resultString = await result.Content.ReadAsStringAsync();
@@ -71,7 +73,9 @@ public class SourceFoldersTests
             true,
             true,
             true,
-            new[] {".jpg", ".png", ".gif", ".webp", ".bmp", ".jpeg", ".tiff", ".tif", ".jfif"});
+            [".jpg", ".png", ".gif", ".webp", ".bmp", ".jpeg", ".tiff", ".tif", ".jfif"],
+            false,
+            null);
         
         var result =await httpClient.PostAsJsonAsync("/collections/source-folders", command);
         var resultString = await result.Content.ReadAsStringAsync();
@@ -86,7 +90,9 @@ public class SourceFoldersTests
             false,
             false,
             false,
-            new[] {".jfif"});
+            [".jfif"],
+            true,
+            "http://example.com/webhook");
         
         await httpClient.PutAsJsonAsync("/collections/source-folders", updateCommand);
         
@@ -121,7 +127,9 @@ public class SourceFoldersTests
             true,
             true,
             true,
-            new[] {".jpg", ".png", ".gif", ".webp", ".bmp", ".jpeg", ".tiff", ".tif", ".jfif"});
+            [".jpg", ".png", ".gif", ".webp", ".bmp", ".jpeg", ".tiff", ".tif", ".jfif"],
+            false,
+            null);
         
         var result =await httpClient.PostAsJsonAsync("/collections/source-folders", command);
         var resultString = await result.Content.ReadAsStringAsync();
@@ -146,7 +154,8 @@ public class SourceFoldersTests
         
         var location = Assembly.GetExecutingAssembly().Location;
 
-        var command1 = new AddSourceFolderCommand(collectionId, Path.Combine(location, "collection", "source 1"), true, true, true, true, new[] {".jpg", ".png", ".gif", ".webp", ".bmp", ".jpeg", ".tiff", ".tif", ".jfif"});
+        var command1 = new AddSourceFolderCommand(collectionId, Path.Combine(location, "collection", "source 1"), true, true, true, true,
+            [".jpg", ".png", ".gif", ".webp", ".bmp", ".jpeg", ".tiff", ".tif", ".jfif"], false, null);
         var command2 = command1 with { Path = Path.Combine(location, "collection", "source 2") };
         var command3 = command1 with { Path = Path.Combine(location, "collection", "source 3") };
         var command4 = command1 with { Path = Path.Combine(location, "collection", "source 4") };